% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adftest.R
\name{adftest}
\alias{adftest}
\title{Augmented Dickey and Fuller (1979) Test}
\usage{
adftest(
  sigl,
  tsrs,
  dpow = -1,
  maxl = NULL,
  crit = "aic",
  smpl = 1000,
  simu = 10000,
  tolr = .Machine$double.eps
)
}
\arguments{
\item{sigl}{Significance level.}

\item{tsrs}{Time series to be tested for a unit root.}

\item{dpow}{Integer specificing power of deterministics to consider (e.g.,
-1: no deterministics, 0: only constant; 1: constant and linear trend).

Default is -1.}

\item{maxl}{Maximum number of lags to consider. Either \code{NULL}
in which case an automated choice is made, or an integer
specifying the maximum number of lags.

Default is \code{NULL}.}

\item{crit}{Information criterion to consider when choosing the
optimal number of lags. Either \code{"aic"} or \code{"bic"}.

Default is \code{"aic"}.}

\item{smpl}{Sample size to consider in case critical values
have to be simulated.}

\item{simu}{Number of repetitions in case critical values have to
be simulated.}

\item{tolr}{Tolerance limit for lookup tables of critical values.}
}
\value{
A list containing:
\itemize{
\item \code{stat}: Test statistic
\item \code{crit}: Critical value
\item \code{rjct}: Logical value indicating the rejection of null hypothesis
}
}
\description{
Performs an augmented Dickey Fuller (1979) (ADF) Test
}
\references{
Dickey, D. A. and Fuller, W. A. (1979). Distribution of the Estimators for
Autoregressive Time Series With a Unit Root. Journal of the American
Statistical Association 74, 427-431.
}
