% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctimcmprtest.R
\name{ctimcmprtest}
\alias{ctimcmprtest}
\title{CT Cointegration Tests for (Systems of) CMPRs Estimated by IM-OLS}
\usage{
ctimcmprtest(sigl, objt, smpl = 1000, simu = 10000, tolr = .Machine$double.eps)
}
\arguments{
\item{sigl}{Significance level}

\item{objt}{Result of a call to \code{imcmpr}}

\item{smpl}{Sample size to consider in case critical values
have to be simulated.}

\item{simu}{Number of repetitions in case critical values have to
be simulated.}

\item{tolr}{Tolerance limit for lookup tables of critical values.}
}
\value{
A list containing:
\itemize{
\item \code{stat}: Test statistic
\item \code{crit}: Critical value
\item \code{rjct}: Logical value indicating the rejection of null hypothesis
}
}
\description{
Performs CT tests for (systems of) CMPRs based a result of a call
to \code{imcmpr}.
}
\references{
Nyblom, J. and Harvey, A. (2000). Tests of Common Stochastic Trends.
Econometric Theory 16, 176-199.

Shin, Y. (1994). A Residual-Based Test of the Null of Cointegration Against
the Alternative of No Cointegration. Econometric Theory 10, 91-115.

Veldhuis, S. and Wagner, M. (2026). Estimation and Inference for
Cointegrated Systems of Multi-Factor Production Functions: Modelling the
Joint Behavior of GDP and Emissions. Mimeo.
}
