% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imcmpr_imp.R
\name{imcmpr_imp}
\alias{imcmpr_imp}
\title{Infeasible (Impracticable) Integrated Modified Estimation of Systems of
Cointegrating Multivariate Polynomial Regressions (IM-SCMPR)}
\usage{
imcmpr_imp(
  yvls,
  zpow,
  xvls,
  clrv,
  long,
  zadd = NULL,
  rmat = NULL,
  rvec = NULL,
  wght = NULL
)
}
\arguments{
\item{yvls}{Matrix of size T x n of integrated variables that cointegrate
with \code{xvls} (see y in the \emph{Details} section); left-hand side of the system
(dependent variables).}

\item{zpow}{Specification of regression setting. Either:
\itemize{
\item An integer, in which case a linear cointegrating
regression with highest polynomial time trend of specified power
is considered (e.g., -1: no deterministic components, 0: constant, 1:
constant and linear trend).
\item Vector of integers of size (m + 1), in which case a cointegrating
polynomial regression (CPR) is considered with the first compoenent
specifying the hightest power of polynomial time trends
and the remaining components specifying the highest powers of the
non-deterministic regressors, i.e., columns of the matrix \code{xvls}.
\item Matrix of size I x (m + 1), in which case a cointegrating multivariate
polynomial regression (CMPR) is considered. The i-th row contains
the multi-index (i_0, i_1, ..., i_m) specifying the powers of the product
z_i = trend^(i_0) * x_1^(i_1) * ... * x_m^(i_m) considered as a
regressor in the regression.
}}

\item{xvls}{Matrix of size T x m of integrated variables that do \emph{not}
cointegrate (see x in the \emph{Details} section).}

\item{clrv}{Matrix of size n x n of the true conditional long-run covariance
matrix Omega_u.v}

\item{long}{Matrix of size (n + m) x (n + m) of the true long-run covariance
matrix Omega}

\item{zadd}{Matrix of size T x q of regressors to be added to the IM
(augmented partial sum) regression (see w in the \emph{Details} section).
Either:
\itemize{
\item \code{NULL}: Empty matrix of size T x 0 (no additional regressors)
\item Matrix of size T x q
}

Default is \code{NULL}.}

\item{rmat}{Restriction matrix of size ((I + m) * n) x g, denoted as D in
the \emph{Details} section, such that vec([cf_mat, cf_aug]') = D * cf_free + d.
Either:
\itemize{
\item \code{NULL}: Identity matrix of size ((I + m) * n) x ((I + m) * n)
\item Matrix of size ((I + m) * n) x g
}

Default is \code{NULL}.}

\item{rvec}{Restriction vector of size (I + m) * n, denoted as d in the
\emph{Details} section, such that vec([cf_mat, cf_aug]') = D * cf_free + d.
Either:
\itemize{
\item \code{NULL}: Zero vector of size (I + m) * n
\item Vector of size (I + m) * n
}

Default is \code{NULL}.}

\item{wght}{Weighting matrix for IM-GLS estimation. Either:
\itemize{
\item \code{NULL}: Identity matrix of size n x n
\item \code{"PaOg1"}: Use Omega_uu^(-1)
\item \code{"PaOg2"}: Use Omega_u.v^(-1)
\item Matrix of size n x n
}

Default is \code{NULL}.}
}
\value{
A list containing:
\itemize{
\item \code{cffs}: Matrix of size n x (I + m) of unrestricted IM-OLS estimates of
[cf_mat, cf_aug].
\item \code{fits}: Matrix of size T x n of fitted values from unrestricted IM-OLS
estimation.
\item \code{rsds}: Matrix of size T x n of residuals from unrestricted IM-OLS
estimation.
\item \code{cvrs}: Matrix of size ((I + m) * n) x ((I + m) * n) giving the
estimated asymptotic covariance matrix of vec(\code{cffs}') for conventional
asymptotic theory.
\item \code{csds}: Matrix of size n x (I + m) giving the estimated asymptotic
standard deviations of \code{cffs} for conventional asymptotic theory.
\item \code{clrv}: Matrix of size n x n of the true conditional long-run
covariance matrix Omega_u.v used for conventional asymptotic theory.
\item \code{rsdm}: Matrix of size T x n of modified IM-OLS residuals used for
fixed-b long-run covariance matrix estimation.
\item \code{cvrb}: Matrix of size ((I + m) * n) x ((I + m) * n) giving the
estimated asymptotic covariance matrix of vec(\code{cffs}') for fixed-b
asymptotic theory.
\item \code{csdb}: Matrix of size n x (I + m) giving the estimated asymptotic
standard deviations of \code{cffs} for fixed-b asymptotic theory.
\item \code{fixb}: Bandwidth-to-sample-size ratio. If \code{band} is  either \code{and} or
\code{nw}, computed usind the first differences of \code{rsdm}. Otherwise,
if \code{band} is either \code{nwt} or an integer in \{1, ..., T\},
the corresponding bandwidth divided by the sample size.
\item \code{cffr}: Matrix of size n x (I + m) giving the restricted IM-GLS
estimates of [cf_mat, cf_aug] with
vec([cf_mat, cf_aug]') = D * cf_free + d.
\item \code{fitr}: Matrix of size T x n of fitted values from restricted IM-GLS
estimation.
\item \code{rsdr}: Matrix of size T x n of residuals from restricted IM-GLS
estimation.
\item \code{cvrr}: Matrix of size ((I + m) * n) x ((I + m) * n) giving the
estimated asymptotic covariance matrix of vec(\code{cffr}') for conventional
asymptotic theory.
\item \code{csdr}: Matrix of size n x (I + m) giving the estimated asymptotic
standard deviations of \code{cffr} for conventional asymptotic theory.
\item \code{cfff}: Vector of size g giving the IM coefficient estimates of cf_free.
\item \code{cvrf}: Matrix of size g x g giving the estimated asymptotic covariance
matrix of \code{cfff} for conventional asymptotic theory.
}
}
\description{
Computes infeasible IM-OLS and IM-GLS estimates for a system of
cointegrating multivariate polynomial regressions.
}
\details{
Model setup:

y = cf_mat * Z + u

Z = [z_1, ..., z_I]'

y = [y_1, ..., y_n]'

with:

z_i = trend^(i_0) * x_1^(i_1) * ... * x_m^(i_m) \cr

IM (augmented partial sum) regression considered:

S^y = cf_mat * S^Z + cf_aug * x + S^u

x = [x_1, ..., x_m]'

where:

vec([cf_mat, cf_aug]') = D * cf_free + d \cr

If \code{zadd} is not \code{NULL}, the IM regression considered becomes:

S^y = cf_mat * S^Z + cf_aug * [w', x']' + S^u

with w corresponding to the regressors specified by \code{zadd}.
}
\note{
The function does not verify asymptotic rank conditions or check that
the restrictions D and d are correctly specified.
It assumes that the restrictions do not involve elements of cf_aug.
}
\references{
Veldhuis, S. and Wagner, M. (2026). Estimation and Inference for
Cointegrated Systems of Multi-Factor Production Functions: Modelling the
Joint Behavior of GDP and Emissions. Mimeo.

Vogelsang, T. J. and Wagner, M. (2014). Integrated Modified OLS Estimation
and Fixed-b Inference for Cointegrated Regressions. Journal of
Econometrics 178, 741-760.

Vogelsang, T. J. and Wagner, M. (2024). Integrated Modified OLS Estimation
and Fixed-b Inference for Cointegrating Multivariate Polynomial
Regressions. IHS Working Paper Series 53.
}
