% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpsstest.R
\name{kpsstest}
\alias{kpsstest}
\title{Kwiatkowski, Phillips, Schmidt and Shin (1992) Test}
\usage{
kpsstest(
  sigl,
  tsrs,
  dpow = NULL,
  krnl = "ba",
  band = "and",
  smpl = 1000,
  simu = 10000,
  tolr = .Machine$double.eps
)
}
\arguments{
\item{sigl}{Significance level.}

\item{tsrs}{Time series to be tested for stationarity.}

\item{dpow}{Integer specificing power of deterministics to consider (e.g.,
\code{-1}: no deterministics, \code{0}: only constant; \code{1}: constant and linear trend).

Default is \code{-1}.}

\item{krnl}{Kernel function used for long-run variance estimation. Either:
\itemize{
\item \code{"tr"}: Truncated
\item \code{"ba"}: Bartlett
\item \code{"pa"}: Parzen
\item \code{"bo"}: Bohman
\item \code{"da"}: Daniell
\item \code{"qs"}: Quadratic Spectral
}

Default is \code{"ba"}.}

\item{band}{Bandwidth specification. Either:
\itemize{
\item Integer in \{1, ..., T\} (explicit bandwidth value)
\item \code{"and"}: Data-dependent rule (Andrews, 1991)
\item \code{"nw"}: Data-dependent rule (Newey & West, 1987)
}

Default is \code{"and"}.}

\item{smpl}{Sample size to consider in case critical values
have to be simulated.}

\item{simu}{Number of repetitions in case critical values have to
be simulated.}

\item{tolr}{Tolerance limit for lookup tables of critical values.}
}
\value{
A list containing:
\itemize{
\item \code{stat}: Test statistic
\item \code{crit}: Critical value
\item \code{rjct}: Logical value indicating the rejection of null hypothesis
}
}
\description{
Performs a KPSS test.
}
\references{
Kwiatkowski, D., Phillips, P. C. B., Schmidt, P. and Shin, Y. (1992).
Testing the Null Hypothesis of Stationarity Against the Alternative of a
Unit Root. Journal of Econometrics 54, 159-178.
}
