% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookupdifu79qntl.R
\name{lookupdifu79qntl}
\alias{lookupdifu79qntl}
\title{Lookup critical values for Dickey and Fuller (1979)}
\usage{
lookupdifu79qntl(prob, dpow, stat, tolr = .Machine$double.eps)
}
\description{
Uses an internal lookup table of Dickey and Fuller (1979) critical values
stored in \code{sysdata.rda}.
}
\keyword{internal}
