% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookupkpss92qntl.R
\name{lookupkpss92qntl}
\alias{lookupkpss92qntl}
\title{Lookup critical values for the
Kwiatkowski, Phillips, Schmidt, Shin (1992) test}
\usage{
lookupkpss92qntl(prob, dpow, tolr = .Machine$double.eps)
}
\description{
Uses an internal lookup table of Kwiatkowski, Phillips, Schmidt, Shin (1992)
critical values stored in \code{sysdata.rda}.
}
\keyword{internal}
