% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookupphou90qntl.R
\name{lookupphou90qntl}
\alias{lookupphou90qntl}
\title{Lookup critical values for Phillips and Ouliaris (1990)}
\usage{
lookupphou90qntl(prob, dpow, stat, xnum, tolr = .Machine$double.eps)
}
\description{
Uses an internal lookup table of Phillips and Ouliaris (1990) critical values
stored in \code{sysdata.rda}.
}
\keyword{internal}
