% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookupsh94qntl.R
\name{lookupsh94qntl}
\alias{lookupsh94qntl}
\title{Lookup critical values for the Shin (1994) test}
\usage{
lookupsh94qntl(prob, xnum, dpow, tolr = .Machine$double.eps)
}
\description{
Uses an internal lookup table of Shin (1994) critical values stored
in \code{sysdata.rda}.
}
\keyword{internal}
