% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookupvewa23qntl.R
\name{lookupvewa23qntl}
\alias{lookupvewa23qntl}
\title{Lookup critical values for Veldhuis and Wagner (????)}
\usage{
lookupvewa23qntl(
  prob,
  dpow,
  xnum,
  stat,
  krnl,
  bfrc,
  tolr = .Machine$double.eps
)
}
\description{
Uses an internal lookup table of Veldhuis and Wagner (????) critical values
stored in \code{sysdata.rda}.
}
\keyword{internal}
