% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookupvewa24ctqntl.R
\name{lookupvewa24ctqntl}
\alias{lookupvewa24ctqntl}
\title{Lookup critical values for Veldhuis and Wagner (????) CT statitics}
\usage{
lookupvewa24ctqntl(prob, ynum, zpow, stat, tolr = .Machine$double.eps)
}
\description{
Uses an internal lookup table of Veldhuis and Wagner (????) CT
critical values stored in \code{sysdata.rda}.
}
\keyword{internal}
