% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookupvewa24fbqntl.R
\name{lookupvewa24fbqntl}
\alias{lookupvewa24fbqntl}
\title{Lookup critical values for Veldhuis and Wanger (...)}
\usage{
lookupvewa24fbqntl(
  prob,
  ynum,
  zpow,
  stat,
  hypo,
  krnl,
  bfrc,
  tolr = .Machine$double.eps
)
}
\description{
Uses an internal lookup table of Veldhuis and Wagner (...) critical values
stored in \code{sysdata.rda}.
}
\keyword{internal}
