% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookupvowa13qntl.R
\name{lookupvowa13qntl}
\alias{lookupvowa13qntl}
\title{Lookup critical values for Vogeland and Wagner (2013)}
\usage{
lookupvowa13qntl(
  prob,
  dpow,
  step,
  stat,
  krnl,
  bfrc,
  tolr = .Machine$double.eps
)
}
\description{
Uses an internal lookup table of Vogeland and Wagner (2013) critical values
stored in \code{sysdata.rda}.
}
\keyword{internal}
