% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookupvowa14qntl.R
\name{lookupvowa14qntl}
\alias{lookupvowa14qntl}
\title{Lookup critical values for Vogelsang and Wagner (2014)}
\usage{
lookupvowa14qntl(
  prob,
  dpow,
  stat,
  xnum,
  hypo,
  krnl,
  bfrc,
  tolr = .Machine$double.eps
)
}
\description{
Uses an internal lookup table of Vogelsang and Wagner (2014) critical values
stored in \code{sysdata.rda}.
}
\keyword{internal}
