% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookupvowa24qntl.R
\name{lookupvowa24qntl}
\alias{lookupvowa24qntl}
\title{Lookup critical values for Vogelsang and Wagner (2024)}
\usage{
lookupvowa24qntl(prob, zpow, hypo, krnl, bfrc, tolr = .Machine$double.eps)
}
\description{
Uses an internal lookup table of Vogelsang and Wagner (2024) critical values
stored in \code{sysdata.rda}.
}
\keyword{internal}
