% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookupwa23qntl.R
\name{lookupwa23qntl}
\alias{lookupwa23qntl}
\title{Lookup critical values for CT and PU cpr}
\usage{
lookupwa23qntl(prob, dpow, xpow, xnum, stat, tolr = .Machine$double.eps)
}
\description{
Uses an internal lookup table of CT and PU cpr critical values
stored in \code{sysdata.rda}.
}
\keyword{internal}
