% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phzfbtest.R
\name{phzfbtest}
\alias{phzfbtest}
\title{Fixed-b Version of Phillips and Perron (1988) Z-Tests}
\usage{
phzfbtest(
  sigl,
  tsrs,
  dpow = -1,
  krnl = "ba",
  band = "and",
  step = 2,
  smpl = 1000,
  simu = 10000,
  tolr = .Machine$double.eps
)
}
\arguments{
\item{sigl}{Significance level.}

\item{tsrs}{Time series to be tested for a unit root.}

\item{dpow}{Integer specificing power of deterministics to consider (e.g.,
\code{-1}: no deterministics, \code{0}: only constant; \code{1}: constant and linear trend).

Default is \code{-1}.}

\item{krnl}{Kernel function used for long-run variance estimation. Either:
\itemize{
\item \code{"tr"}: Truncated
\item \code{"ba"}: Bartlett
\item \code{"pa"}: Parzen
\item \code{"bo"}: Bohman
\item \code{"da"}: Daniell
\item \code{"qs"}: Quadratic Spectral
}

Default is \code{"ba"}.}

\item{band}{Bandwidth specification. Either:
\itemize{
\item Integer in \{1, ..., T\} (explicit bandwidth value)
\item \code{"and"}: Data-dependent rule (Andrews, 1991)
\item \code{"nw"}: Data-dependent rule (Newey & West, 1987)
}

Default is \code{"and"}.}

\item{step}{Integer in \{\code{1}, \code{2}\} specifying whether one- or two-step
detrending should be applied. Either:
\itemize{
\item \code{1}: One-step detrending
\item \code{2}: Two-step detrending
}

Default is \code{2}.}

\item{smpl}{Sample size to consider in case critical values
have to be simulated.}

\item{simu}{Number of repetitions in case critical values have to
be simulated.}

\item{tolr}{Tolerance limit for lookup tables of critical values.}
}
\value{
A list containing:
\itemize{
\item \code{stat}: Test statistic
\item \code{crit}: Critical value
\item \code{rjct}: Logical value indicating the rejection of null hypothesis
}
}
\description{
Performs the fixed-b version of the Phillips and Perrson (1988) Z-tests
provided by Vogelsang and Wagner (2013)
}
\references{
Phillips, P. C. B. (1987). Time Series Regression with a Unit Root.
Econometrica 55, 277-301.

Phillips, P. C. B. and Perron, P. (1988). Testing for a Unit Root in Time
Series Regression. Biometrika 75, 335-346.

Vogelsang, T. J. and Wagner, M. (2013). A Fixed-b Perspective on the
Phillips-Perron Unit Root Tests. Econometric Theory 29, 609-628.
}
