% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phztest.R
\name{phztest}
\alias{phztest}
\title{Phillips and Perron (1988) Z-Tests}
\usage{
phztest(
  sigl,
  tsrs,
  dpow = -1,
  krnl = "ba",
  band = "and",
  step = 2,
  smpl = 1000,
  simu = 10000,
  tolr = .Machine$double.eps
)
}
\arguments{
\item{sigl}{Significance level.}

\item{tsrs}{Time series to be tested for a unit root.}

\item{dpow}{Integer specificing power of deterministics to consider (e.g.,
-1: no deterministics, 0: only constant; 1: constant and linear trend).

Default is -1.}

\item{krnl}{Kernel function used for long-run variance estimation. Either:
\itemize{
\item \code{"tr"}: Truncated
\item \code{"ba"}: Bartlett
\item \code{"pa"}: Parzen
\item \code{"bo"}: Bohman
\item \code{"da"}: Daniell
\item \code{"qs"}: Quadratic Spectral
}

Default is \code{"ba"}.}

\item{band}{Bandwidth specification. Either:
\itemize{
\item \code{"and"}: Data-dependent rule (Andrews, 1991)
\item \code{"nw"}: Data-dependent rule (Newey & West, 1987)
}

Default is \code{"and"}.}

\item{step}{Integer, either 1 or 2. Specifys whether one- or two-step
detrending is applied.}

\item{smpl}{Sample size to consider in case critical values
have to be simulated.}

\item{simu}{Number of repetitions in case critical values have to
be simulated.}

\item{tolr}{Tolerance limit for lookup tables of critical values.}
}
\value{
A list containing:
\itemize{
\item \code{stat}: Test statistic
\item \code{crit}: Critical value
\item \code{rjct}: Logical value indicating the rejection of null hypothesis
}
}
\description{
Performs Phillips and Perrson (1988) Z-tests
}
\references{
Phillips, P. C. B. (1987). Time Series Regression with a Unit Root.
Econometrica 55, 277-301.

Phillips, P. C. B. and Perron, P. (1988). Testing for a Unit Root in Time
Series Regression. Biometrika 75, 335-346.
}
