% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/povrcmprtest.R
\name{povrcmprtest}
\alias{povrcmprtest}
\title{Phillips-Ouliaris (1990) Variance Ratio Test}
\usage{
povrcmprtest(
  sigl,
  yvls,
  zpow,
  xvls,
  krnl = "ba",
  band = "and",
  smpl = 1000,
  simu = 10000,
  tolr = .Machine$double.eps
)
}
\arguments{
\item{sigl}{Significance level.}

\item{yvls}{Values of the integrated dependent variable.}

\item{zpow}{Specification of regression setting. Either:
\itemize{
\item An integer, in which case a linear
regression with highest polynomial time trend of specified power
is considered (e.g., -1: no deterministic components, 0: constant, 1:
constant and linear trend).
\item Vector of integers of size, in which case a
polynomial regression (PR) is considered with the first compoenent
specifying the hightest power of polynomial time trends
and the remaining components specifying the highest powers of the
non-deterministic regressors, i.e., columns of the matrix \code{xvls}.
\item A matrix, in which case a multivariate polynomial
regression (MPR) is considered. Each row contains a multi-index
(i_0, i_1, ..., i_m) specifying the powers of the product
z_i = trend^(i_0) * x_1^(i_1) * ... * x_m^(i_m) considered as a
regressor in the regression. Here x_1, ..., x_m denote the columns
of the matrix \code{xvls}.
}}

\item{xvls}{Matrix of integrated variables that do \emph{not} cointegrate.}

\item{krnl}{Kernel function used for long-run variance estimation. Either:
\itemize{
\item \code{"tr"}: Truncated
\item \code{"ba"}: Bartlett
\item \code{"pa"}: Parzen
\item \code{"bo"}: Bohman
\item \code{"da"}: Daniell
\item \code{"qs"}: Quadratic Spectral
}

Default is \code{"ba"}.}

\item{band}{Bandwidth specification. Either:
\itemize{
\item \code{"and"}: Data-dependent rule (Andrews, 1991)
\item \code{"nw"}: Data-dependent rule (Newey & West, 1987)
}

Default is \code{"and"}.}

\item{smpl}{Sample size to consider in case critical values
have to be simulated.}

\item{simu}{Number of repetitions in case critical values have to
be simulated.}

\item{tolr}{Tolerance limit for lookup tables of critical values.}
}
\value{
A list containing:
\itemize{
\item \code{stat}: Test statistic
\item \code{crit}: Critical value
\item \code{rjct}: Logical value indicating the rejection of null hypothesis
}
}
\description{
Performs a non-cointegration Phillips-Ouliaris (1990) Variance Ratio (PU)
Test, extended to the cointegrating polynomial regressions (CPR) setting
by Wagner (2023) and to the cointegrating multivariate polynomial
regressions (CMPR) Setting by Stypka and Wagner (2024)
}
\references{
Phillips, P. C. B. and Ouliaris, S. (1990). Asymptotic Properties of
Residual Based Tests for Cointegration. Econometrica 58, 165-193.

Wagner, M. (2023). Residual-Based Cointegration and Non-Cointegration
Tests for Cointegrating Polynomial Regressions. Empirical Economics 65,
1-31.
}
