% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poztest.R
\name{poztest}
\alias{poztest}
\title{Phillips-Ouliaris (1990) Version of Phillips and Perron (1988) Z-Tests}
\usage{
poztest(
  sigl,
  yvls,
  xvls,
  dpow = -1,
  krnl = "ba",
  band = "and",
  smpl = 1000,
  simu = 10000,
  tolr = .Machine$double.eps
)
}
\arguments{
\item{sigl}{Significance level.}

\item{yvls}{Values of the integrated dependent variable.}

\item{xvls}{Matrix of integrated regressors that do \emph{not} cointegrate.}

\item{dpow}{Integer specificing power of deterministics to consider (e.g.,
-1: no deterministics, 0: only constant; 1: constant and linear trend).

Default is -1.}

\item{krnl}{Kernel function used for long-run variance estimation. Either:
\itemize{
\item \code{"tr"}: Truncated
\item \code{"ba"}: Bartlett
\item \code{"pa"}: Parzen
\item \code{"bo"}: Bohman
\item \code{"da"}: Daniell
\item \code{"qs"}: Quadratic Spectral
}

Default is \code{"ba"}.}

\item{band}{Bandwidth specification. Either:
\itemize{
\item \code{"and"}: Data-dependent rule (Andrews, 1991)
\item \code{"nw"}: Data-dependent rule (Newey & West, 1987)
}

Default is \code{"and"}.}

\item{smpl}{Sample size to consider in case critical values
have to be simulated.}

\item{simu}{Number of repetitions in case critical values have to
be simulated.}

\item{tolr}{Tolerance limit for lookup tables of critical values.}
}
\value{
A list containing:
\itemize{
\item \code{stat}: Test statistics
\item \code{crit}: Critical values
\item \code{rjct}: Logical values indicating the rejection of null hypothesis
}
}
\description{
Performs non-cointegration Phillips-Ouliaris (1990) Z-tests
}
\references{
Phillips, P. C. B. and Ouliaris, S. (1990). Asymptotic Properties of
Residual Based Tests for Cointegration. Econometrica 58, 165-193.

Phillips, P. C. B. and Perron, P. (1988). Testing for a Unit Root in Time
Series Regression. Biometrika 75, 335-346.
}
