% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinimcmprtest.R
\name{shinimcmprtest}
\alias{shinimcmprtest}
\title{IM-OLS Based Shin (1994) Cointegration Test}
\usage{
shinimcmprtest(
  sigl,
  yvls,
  zpow,
  xvls,
  zadd = NULL,
  krnl = "ba",
  band = "and",
  dtrn = NULL,
  step = 2,
  smpl = 1000,
  simu = 10000,
  tolr = .Machine$double.eps
)
}
\arguments{
\item{sigl}{Significance level.}

\item{yvls}{Values of the integrated dependent variable.}

\item{zpow}{Specification of regression setting. Either:
\itemize{
\item An integer, in which case a linear
regression with highest polynomial time trend of specified power
is considered (e.g., -1: no deterministic components, 0: constant, 1:
constant and linear trend).
\item Vector of integers of size, in which case a
polynomial regression (PR) is considered with the first compoenent
specifying the hightest power of polynomial time trends
and the remaining components specifying the highest powers of the
non-deterministic regressors, i.e., columns of the matrix \code{xvls}.
\item A matrix, in which case a multivariate polynomial
regression (MPR) is considered. Each row contains a multi-index
(i_0, i_1, ..., i_m) specifying the powers of the product
z_i = trend^(i_0) * x_1^(i_1) * ... * x_m^(i_m) considered as a
regressor in the regression. Here x_1, ..., x_m denote the columns
of the matrix \code{xvls}.
}}

\item{xvls}{Matrix of integrated variables that do \emph{not} cointegrate.}

\item{zadd}{Matrix of regressors to be added to the IM (augmented partial
sum) regression. Either \code{NULL}, in which case no additional regressors are
considered or a matrix of additional regressors.

Default is \code{NULL}.}

\item{krnl}{Kernel function used for long-run variance estimation. Either:
\itemize{
\item \code{"tr"}: Truncated
\item \code{"ba"}: Bartlett
\item \code{"pa"}: Parzen
\item \code{"bo"}: Bohman
\item \code{"da"}: Daniell
\item \code{"qs"}: Quadratic Spectral
}

Default is \code{"ba"}.}

\item{band}{Bandwidth specification. Either:
\itemize{
\item \code{"and"}: Data-dependent rule (Andrews, 1991)
\item \code{"nw"}: Data-dependent rule (Newey & West, 1987)
}

Default is \code{"and"}.}

\item{dtrn}{Detrending of \code{xvls} to be applied in long-run covariance matrix
estimation under conventional asymptotic theory. Either:
\itemize{
\item \code{NULL}: No detrening is applied
\item Matrix used as detrending regressors
}

Default is \code{NULL}.}

\item{step}{Integer in \{\code{1}, \code{2}\} specifying whether one- or two-step
detrending in long-run covariance matrix estimation under conventional
asymptotic theory is applied (only used, if \code{dtrn} is not \code{NULL}). Either:
\itemize{
\item \code{1}: One-step detrending
\item \code{2}: Two-step detrending
}}

\item{smpl}{Sample size to consider in case critical values
have to be simulated.}

\item{simu}{Number of repetitions in case critical values have to
be simulated.}

\item{tolr}{Tolerance limit for lookup tables of critical values.}
}
\value{
A list containing:
\itemize{
\item \code{stat}: Test statistic
\item \code{crit}: Critical value
\item \code{rjct}: Logical value indicating the rejection of null hypothesis
}
}
\description{
Performs an IM-OLS based Shin (1994) Cointegration Test (CT), provided
(implicitly) by Wagner and Wied (2017), extended to the cointegrating
multivariate polynomial regressions (CPR) setting by Veldhuis and
Wagner (2026)
}
\references{
Shin, Y. (1994). A Residual-Based Test of the Null of Cointegration Against
the Alternative of No Cointegration. Econometric Theory 10, 91-115.

Veldhuis, S. and Wagner, M. (2026). Estimation and Inference for
Cointegrated Systems of Multi-Factor Production Functions: Modelling the
Joint Behavior of GDP and Emissions. Mimeo.
}
