% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timcmprfbtest.R
\name{timcmprfbtest}
\alias{timcmprfbtest}
\title{Fixed-b t-Type Tests for CMPRs Estimated by IM-OLS}
\usage{
timcmprfbtest(
  sigl,
  objt,
  rvec,
  rval,
  dirc = 0,
  smpl = 1000,
  simu = 10000,
  tolr = .Machine$double.eps
)
}
\arguments{
\item{sigl}{Significance level}

\item{objt}{Result of a call to \code{imcmpr} with \code{ncol(yvls) == 1}}

\item{rvec}{Restriction vector}

\item{rval}{Restriction value}

\item{dirc}{Integer specifying the test direction: -1 for left-tailed,
1 for right-tailed, or 0 for two-sided.}

\item{smpl}{Sample size to consider in case critical values
have to be simulated.}

\item{simu}{Number of repetitions in case critical values have to
be simulated.}

\item{tolr}{Tolerance limit for lookup tables of critical values.}
}
\value{
A list containing:
\itemize{
\item \code{stat}: Test statistic
\item \code{crit}: Critical value
\item \code{rjct}: Logical value indicating the rejection of null hypothesis
}
}
\description{
Performs a fixed-b t-type test based a result of a call to \code{imcmpr}.
}
\references{
Veldhuis, S. and Wagner, M. (2026). Estimation and Inference for
Cointegrated Systems of Multi-Factor Production Functions: Modelling the
Joint Behavior of GDP and Emissions. Mimeo.

Vogelsang, T. J. and Wagner, M. (2014). Integrated Modified OLS Estimation
and Fixed-b Inference for Cointegrated Regressions. Journal of
Econometrics 178, 741-760.

Vogelsang, T. J. and Wagner, M. (2024). Integrated Modified OLS Estimation
and Fixed-b Inference for Cointegrating Multivariate Polynomial
Regressions. IHS Working Paper Series 53.
}
