% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlstest.R
\name{tlstest}
\alias{tlstest}
\title{Conventional t-Type Tests for CMPRs Estimated by
Integrated Modified Least Squares}
\usage{
tlstest(sigl, objt, rvec, rval, rstr = FALSE, dirc = 0)
}
\arguments{
\item{sigl}{Significance level}

\item{objt}{Result of a call to \code{imcmpr}}

\item{rvec}{Restriction vector}

\item{rval}{Restriction value}

\item{rstr}{Logical. Should the Wald test be formed based for the free
parameters of the restricted estimator?}

\item{dirc}{Integer specifying the test direction: -1 for left-tailed,
1 for right-tailed, or 0 for two-sided.}
}
\value{
A list containing:
\itemize{
\item \code{stat}: Test statistic
\item \code{crit}: Critical value
\item \code{rjct}: Logical value indicating the rejection of null hypothesis
}
}
\description{
Performs a conventional t-type test based a result of a call to \code{imcmpr}.
}
\references{
Veldhuis, S. and Wagner, M. (2026). Estimation and Inference for
Cointegrated Systems of Multi-Factor Production Functions: Modelling the
Joint Behavior of GDP and Emissions. Mimeo.

Vogelsang, T. J. and Wagner, M. (2014). Integrated Modified OLS Estimation
and Fixed-b Inference for Cointegrated Regressions. Journal of
Econometrics 178, 741-760.

Vogelsang, T. J. and Wagner, M. (2024). Integrated Modified OLS Estimation
and Fixed-b Inference for Cointegrating Multivariate Polynomial
Regressions. IHS Working Paper Series 53.
}
