% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waldlstest.R
\name{waldlstest}
\alias{waldlstest}
\title{Conventional Wald-Type Tests for (Systems of) CMPRs Estimated by
Integrated Modified Least Squares}
\usage{
waldlstest(sigl, objt, rmat, rvec, rstr = FALSE)
}
\arguments{
\item{sigl}{Significance level}

\item{objt}{Result of a call to \code{imls}}

\item{rmat}{Restriction matrix}

\item{rvec}{Restriction vector}

\item{rstr}{Logical. Should the Wald test be formed based for the free
parameters of the restricted estimator?}
}
\value{
A list containing:
\itemize{
\item \code{stat}: Test statistic
\item \code{crit}: Critical value
\item \code{rjct}: Logical value indicating the rejection of null hypothesis
}
}
\description{
Performs a conventional Wald-type test based a result of a call to \code{imcmpr}.
}
\references{
Veldhuis, S. and Wagner, M. (2026). Estimation and Inference for
Cointegrated Systems of Multi-Factor Production Functions: Modelling the
Joint Behavior of GDP and Emissions. Mimeo.

Vogelsang, T. J. and Wagner, M. (2014). Integrated Modified OLS Estimation
and Fixed-b Inference for Cointegrated Regressions. Journal of
Econometrics 178, 741-760.

Vogelsang, T. J. and Wagner, M. (2024). Integrated Modified OLS Estimation
and Fixed-b Inference for Cointegrating Multivariate Polynomial
Regressions. IHS Working Paper Series 53.
}
